requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
plotly>=5.18.0
pandas>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
